First two functions
I completed the functionality for two parts of the website. These two are the available sports option to get sport Ids, and I also completed the events by sport id and date functionality.

The available sports shows the sports available through the API and the sport Id linked to each one.

The event by sport id and date shows the events for the selected sport on the selected day. This will also provide the event id for that particular event.