Added event details function
I added the event details functionality.

This will now provide details for the event using the event Id. This will show the name of the event which also includes the schedule. It will show the score which will be updated as the game is played. The last thing it will show is the winner after the game has been finished.